Worked on threading, utils, and more!
I added some cool methods to the utilities class, added the abilities
for servers and clients to run on their own thread without an external
one having to be created, added proper order channel handling for
EncapsulatedPackets for receiving and sending, and started work on a
client! I also did some tidying up because the class structure was
messy. ClientSession and ServerSession both new extend RakNetSession, an
abstract class. I did this because ClientSession and ServerSession are
so similar in their class structure. When I can get a RakNetClient fully
functioning, I believe I will be able to release JRakNet v1.0!