Worked on protocols
I found out some of my packets had wrong variable names and/or were
encoded and decoded completely wrong! Luckily RakNet seems to be loose
with it's protocol. I still fixed the issues and updated the protools.
The only packet that was really missing stuff was UnconnectedPing, there
was apparently a field where it also sent it's clientId as a long. I've
also added a feature where in RakNetOptions it is possible to have
exceptions completely ignored by the Netty handler. I'm not sure if it
works yet though, I have to find out a good way to test it.