Removed IgnoredExceptions concept
I decided to remove the IgnoredHandlerExceptions idea as I consider it
bad practice and it didn't work. Also, there seems to currently be a
glitch with RakNet protocol 8 where ID_UNCONNECTED_PONG (0x1C) where the
identifier is not included (Not even the short for the length!) so an
exception is thrown. I fixed this by adding a check in UnconnectedPing
and UnconnectedLegacyPing.