DID MORE COOL STUFF
I added actual Incorrect protocol packets, turns out they were changed
to ID 0x19 right under our noses! I also found out about 0x14, a "Server
full" packet thanks to MiNET documentation. Since it seems to be RakNet
protocol implementation, RakNetOptions now has an serverMaxConnectins
option which defaults to 10. I might make it 0 however to force people
to set it and default player counts might drastically change depending
on the game. Also, player count is dependent on all clientsessions. If
the client is at least in stage CONNECTING_1, it is first come first
served. So that client gets a spot reserved for them. If the maximum
session count is reached, when a 0x05 is received the server responds
with 0x14 and the session is removed. Also RakNetClient's now throw more
exceptions when logging in. Before, it would just throw everything to
the handler. However, now it will throw it directly in the method using
the magic of ArrayList's. If a packet such as SERVER_FULL or
INCOMPATIBLE_PROTOCOL is received it adds the exception to the list and
the connection loop breaks and throws the first exception it finds in
that list. There are more things I added and don't remember. I think 1.5
will be called "The useful update" as it makes use RakNet protocol 8
packets more :P