Worked on packets and exceptions
I added a RakNetClientException, which is an exception that is thrown
only by the client. I also added more exceptions like
ConnectionBannedException and ServerFullException. All the old client
exceptions were modified to extend RakNetClientException. I also added a
new package in the protocol package called "unknown" which has all the
unknown packets that are seen to affect the client in someway. They all
contain the packet ID, the time they were discovered, and notes about
them. In the exceptionCaught method for RakNetServerHandler, it is
checked to make sure the actual session is not null. I also added
UnconnectedConnectionBanned, which is the packet which causes
RakNetCilent to throw ConnectionBannedException. If a client is blocked
on the server, that packet is sent to them directly through the
ChannelHandlerContext without creating a new ClientSession to send it.