Worked on some stuff
I saw some design inconsistencies and a bug where the server and the
client would not make sure the MTU set in the options was the correct
amount. I'm considering getting rid of RakNetOptions entirely and make
it so devolopers have to set everything directly in the constructor to
make it easier for the internal design.