Worked one lotso stuff
I've been trying to fix latency detection, and it's getting there! It
seems like RakNet 8 CONNECTED_PONG does not have any timestamps in them,
as any official server sending one causes the handler to glitch out and
disconnect. It seems because of this I am unable to add latency
detection for the server, but if my theory is correct where it is only
the ID now then I will update the packet formats and just update the
latency based on System.currentTimeMillis() when a response is received.

I did however find out the ConnectedCloseConnection is TWO WAY! If sent
from the server, it will cause the client to stop trying to connect.
Very useful! Whenever removing a session that packet is sent, and
whenever an address is blocked, if a session with that address exists,
it will remove the session from the handler as well (thus causing that
send). If any packets attempt to send to the server and the address is
blocked, an ConnectedCloseConnection and UnconnectedConnectionBanned are
sent now :)