Added MessageIdentifiers
I decided to move all the packets to a class called MessagedIdentifiers,
RakNet is still an interface but contains only things like timeouts and
what not. The relibility types are now in an interface in the
Reliability class. If a class uses alot of Reliability references a
developer can implement Reliability.INTERFACE