Updated hook and client disconnection
I updated HANDLER_EXCEPTION_OCCURED so it now has a 3rd argument which
is possible to be a RakNetSession, if one was not found it will simply
be null. HANDLER_EXCEPTION_OCCURED will also now be executed before the
session is removed if one was found. Also, when sessions are removed
from the server handler they are no longer sent a
ConnectedConnectionClose packet, those are now only sent when
server.kickClient() is called. Version bumped to 1.6.4